Latest version of Frontend
The current build of the frontend
Improved handling of attempts to submit books without specifiying a Shelf Location.
Added explinations to BookAdder and StackAdder to help new users.
Changed tab names for BookAdder and StackAdder to be more intuitive.
Cleaned up appearance of forms a bit.
Changed the Favicon to books.